[refactor] entity package 변경
