+++ data/src/main/java/com/prac/data/repository/RepoRepository.kt
@@ -1,6 +1,6 @@
 package com.prac.data.repository
 
-import com.prac.data.di.entity.RepoEntity
+import com.prac.data.entity.RepoEntity
 
 interface RepoRepository {
     suspend fun getRepositories(
